Abstract up subscript(range:) to SpanningContainer
